Refactored SimpleBottomNavigation to two buttons and nav files.
